create README.md
The warburg effect is an important actor in cancer cellular reprogramation. Being able to target genes reponsable for this reprogramming would contitute an interesting therapeutical target. 

In  this Jupyter notebook you will find how different model were used and optimized in order to get a better prediction of warburg genes.